requirements.txt
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
monotonic==1.5
pymongo==3.7.2
python-engineio==3.4.4
python-socketio==4.0.0
six==1.12.0
Werkzeug==0.14.1
